Refactor figure 1 into component
