+++ app/work/reviewed/Stem.tsx
@@ -1,0 +1,13 @@
+const Stem = () => (
+  <svg
+    width="3"
+    height="28"
+    viewBox="0 0 3 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M1.5 0L1.5 28" stroke="black" strokeWidth="2" />
+  </svg>
+);
+
+export default Stem;
